+++ src/main/java/software/bernie/example/registry/BlockRegistry.java
@@ -2,7 +2,6 @@ package software.bernie.example.registry;
 
 import software.bernie.example.block.BotariumBlock;
 import software.bernie.example.block.FertilizerBlock;
-import software.bernie.geckolib3.util.RegistryUtils;
 
 public class BlockRegistry {
     public static final BotariumBlock BOTARIUM_BLOCK = RegistryUtils.register("botariumblock", new BotariumBlock());
