+++ src/main/java/software/bernie/example/registry/TileRegistry.java
@@ -3,7 +3,6 @@ package software.bernie.example.registry;
 import net.minecraft.block.entity.BlockEntityType;
 import software.bernie.example.block.tile.BotariumTileEntity;
 import software.bernie.example.block.tile.FertilizerTileEntity;
-import software.bernie.geckolib3.util.RegistryUtils;
 
 public class TileRegistry {
     public static final BlockEntityType<BotariumTileEntity> BOTARIUM_TILE = RegistryUtils.registerBlockEntity("botariumtile", BlockEntityType.Builder.create(BotariumTileEntity::new, BlockRegistry.BOTARIUM_BLOCK));
